Coin-change-recursive approach
This solution was accepted in Leetcode. But it gives me Time exceeded error. Hence, I will be implementing the same using Dynamic programming. 
Problem faced: If I create the helper function outside the coinChange function, then it throws an error stating 'helper is not defined'. The program runs only if I give helper function inside the conChange function block and then call it on the return statement of the outer function block. Is it a correct way to implement. Is there any better ways to call helper function from main function?